+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - add-service: Autocreate the config file if missing
  - Better logs when a service fails to start
  - Add grace period, `start` waits 500ms to ensure the process doesn't fail on startup
+ - Better support when commands are called with multiple service names.
 
 # 0.9.0
  - Show observed logs when wait-for-log fails.